refactor: add a module for field types
